LayoutTests/fast/canvas/context-globalCompositeOperation-null-expected.txt

Test the behavior when assigning null to context.globalCompositeOperation

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context.globalCompositeOperation is "source-over"
PASS context.globalCompositeOperation = 'copy' did not throw exception.
PASS context.globalCompositeOperation is "copy"
PASS context.globalCompositeOperation = null did not throw exception.
PASS context.globalCompositeOperation is "copy"
PASS context.globalCompositeOperation = '' did not throw exception.
PASS context.globalCompositeOperation is "copy"
PASS successfullyParsed is true

TEST COMPLETE
